Added feature to display on bottom, #1
Display on bottom works but with a cons. The bottom is the last line of
the buffer which may not be the bottom of the window, If there is only a
few line in the document it will display on the last line. In future if
I find a fix for that I will do it.